+++ src/archive/archive-reader.test.ts
@@ -3,11 +3,11 @@ import { InMemoryAsyncUint8Array } from '../common/async-uint8-array';
 import { AsyncUint8ArrayIterator } from '../common/async-uint8-array-iterator';
 import { Constants } from '../common/constants';
 import { TarUtility } from '../common/tar-utility';
+import { PaxTarHeader, PaxTarHeaderAttributes } from '../header/pax/pax-tar-header';
+import { PaxTarHeaderKey } from '../header/pax/pax-tar-header-key';
 import { TarHeader } from '../header/tar-header';
 import { TarHeaderLike } from '../header/tar-header-like';
 import { TarHeaderLinkIndicatorType } from '../header/tar-header-link-indicator-type';
-import { PaxTarHeader, PaxTarHeaderAttributes } from '../pax/pax-tar-header';
-import { PaxTarHeaderKey } from '../pax/pax-tar-header-key';
 import { tarballSampleBase64 as PAX_tarballSampleBase64, totalFileCount as PAX_totalFileCount } from '../test/generated/pax-header-test-content';
 import { base64ToUint8Array, range } from '../test/test-util';
 import { ArchiveReader, ArchiveReadError } from './archive-reader';
